+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - Fix display-nonblocking example
+- Fix timer for LED-display (GreyscaleImage with a empty Row did not work)
 
 ## [0.12.0] - 2021-11-10